move smsbillables generator to tests
