+++ corehq/apps/smsbillables/tests/test_usage_fees.py
@@ -5,14 +5,13 @@ from django.test import TestCase
 
 from corehq.apps.accounting.tests.generator import init_default_currency
 from corehq.apps.sms.models import SQLMobileBackend
-from corehq.apps.smsbillables import generator
 from corehq.apps.smsbillables.models import (
     SmsBillable,
     SmsGatewayFee,
     SmsUsageFee,
     SmsUsageFeeCriteria,
 )
-from corehq.apps.smsbillables.tests import FakeTwilioMessageFactory
+from corehq.apps.smsbillables.tests import FakeTwilioMessageFactory, generator
 from corehq.messaging.smsbackends.twilio.models import SQLTwilioBackend
